proposed change for timeout in http:send-request()
Right now, the timeout only sets the connect timeout that is active
while a connection is established, but not the read timeout while
waiting for a response.

The EXPath HTTP Client Module specification does not clearly state the
semantics of the parameter "timeout", but if the read timeout is not
considered in the implementation, a send-request() may hang forever
even though a timeout parameter has been provided.